Splitting layer map generation
