+++ Assets/Scripts/Terrain/PathGraph/CellularAutomata/ILayerGenerator.cs
@@ -1,0 +1,10 @@
+﻿using Unity.Collections;
+using Unity.Mathematics;
+
+namespace Terrain.PathGraph.CellularAutomata
+{
+    public interface ILayerGenerator
+    {
+        public void GenerateLayer(NativeArray<byte> baseMap, Layer[] layers, int2 mapSize);
+    }
+}
