--- Assets/Scripts/Terrain/PathGraph/CellularAutomata/InitialMapJobs.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 406e8226a8a14575a08d121b63d1a4a5
-timeCreated: 1689366186
